+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.39.19 (October 23, 2019)
+
+- fix: CLI still fetch remote assets while `releaseDir` was set #302
+- fix: `host` is require by latest workshop config #304
+
 ## 0.39.18 (October 23, 2019)
 
 - validate if the chain config folder is complete #282